Allow to migrate all migrations with one "mark_migrated" call
Added a special value "all" for the "version" argument of the "mark_migrated" command.
It allows to mark all migrations found (based on the other arguments) as migrated in one call.
Everything is done in one transaction. If one the marking as migrated operations raises an exception, the entire process is cancelled and the transaction rollbacked.