Add precision handling to the snapshot template
By adding precision to the wantedOptions, we can handle it in the template itself as it won't be
dropped out.

If precision is a non-empty value, we need to swap around limit, precision and scale to use the Phinx
usages of these terms with regard to decimal fields so that the migration is rendered correctly.  Otherwise
unset the value so we don't have precision appearing on every field.