Updated function calls
Phinx v0.7.0 changed `getMigrationPath()` to `getMigrationPaths()`
and `getSeedPath()` to `getSeedPaths()`.  The older methods were
returning a single path, while the newer ones are returning an
array of locations.

The easist workaround is to just `array_pop()` the path from the
array and use that.

The only annoying bit is that `array_pop` expects a reference to the
array, so the method names don't really work.  Hence, all the extra
temporary variables like `$migrationPaths` and `$seedPaths`.