Add error message when using runMany() and skip
Applying `skip` to multiple profiles when using `runMany()` creates
a problem that I don't think is solvable.

The problem boils down to skip is able to exclude some or all tables
from a migration set (plugin or application). When we're attempting to
reset both table schema and phinxlog state we can only operate at the
connection level. Because both your plugin and application are using the
same connection the queries we need to use to find which tables need to
be dropped and which phinxlogs need to be reset pollute each other and
overlap.

Because we don't have enough metadata to know which tables map to which
phinxlog tables, and which migrations within those migration logs we can
only operate on schema and phinxlogs in a coarse manner.

Refs #603