Import Db/Table package and get tests passing
Import the Db/Table package from phinx and get the tests passing.

My plan here is to import all of the 'data transfer' objects from phinx
and get them referencing each other. Once that is complete, we'll need
to import the Plan and Migration wrappers.

This will form the core of the API compatibility with phinx. The adapter
layer can be replaced by a mix of Cake's Database package and some more
dialect style platform wrappers.

Finally we'll need to provide the same CLI interface that migrations has
always given.