Get started on PhinxAdapter
Get the rough draft of a shim adapter that will bridge between the new
migrations backend and phinx's adapter interfaces. The new backend will
create an instance of this adapter when running migrations and seeds.
This will allow migrations to stay compatible with phinx until
a future major release. I want the new backend to be opt-in initially
so that we can release it, test and iterate and stabilize it when it is
ready without risking broken releases.

I've run into a blocker with phinx though that I need to patch first.
Several DTO classes have uninitialized properties that make using their
public interfaces unsafe.