Make builtin the default migrations backend
The 4.3 branch has been out for a few months and there haven't been any
showstopping bugs reported yet. Changing the default backend to builtin
and retaining the phinx backend as an escape hatch until a fix can be
released seems like a decent compromise on moving forwards and retaining
compatibility.